Revert "Log unhandled exceptions"
This reverts commit b0339f4a8f29df9ab4b65b94cf8f6fa0d45c51e0.

BaseException catches regular SystemExit used to terminate the script when using parameters to generate the JSON file and so on.

Catching and logging unhandled exeptions needs to be implemented differently or at least require further changes to how the script exits when it does it normally.

For the time being the previous attempt at logging unhandled expections is reverted.

Closes #13

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>