fix: move logging earlier in main()
As a result of PR #10 we now try to log unhandled exceptions using the log function. However, whether logging is enabled or not is determined at a point in main where exceptions could have occurred leading to two exceptions, one to do with the original event and one to do with logging not being defined yet.

This change brings logging to an earlier point to try to avoid that.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>